adapting new format filter

+++ Domain/ApartmentRank.Domain/ValueObjects/Filter.cs
@@ -5,7 +5,7 @@
         public string operation { get; set; }
         public string propertyType { get; set; }
         public string center { get; set; }
-        public int distance { get; set; }
+        public double distance { get; set; }
         public int maxPrice { get; set; }
         public bool studio { get; set; }
         public int bedrooms { get; set; }
